Remove unused components related to back-burnered forms system rewrite
