+++ packages/lesswrong/components/form-components/PostsListEditor.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { makeSortableListComponent } from '../forms/sortableList';
+import { makeSortableListComponent } from './sortableList';
 import { registerComponent, Components } from '../../lib/vulcan-lib';
 import * as _ from 'underscore';
